requirements.txt
uv
streamlit
python-dotenv
langchain
langchain-google-genai
langgraph
langchain_community
alpha_vantage
newsapi-python
pydantic
flask
flask-cors